Phase 0: float Pipeline + retune handshake
- IPipelineHandler::processBlock теперь принимает const float* нормированный к [-1,1]
- Единая int16→float конвертация в RxWorker на границе с железом
- Все handlers (FFT, Demod, Bandpass, RawFile, Classifier) переведены на float
- RawFileHandler пишет .cf32 вместо .cs16
- Тесты FFT/FM/AM обновлены под float API

Retune handshake (UI ↔ worker):
- LimeDevice::setFrequency(ch, hz) на стримящем канале паркует RxWorker
  через checkPauseForRetune, выполняет LMS_StopStream/SetLO/StartStream
  на UI-потоке, затем отпускает worker
- Новый сигнал IDevice::retuned(ch, hz), соединён DirectConnection из
  RxController → Pipeline::notifyRetune → handler onRetune()
- Handlers с накопленным DSP state (FIR delay, NCO phase, WAV decimation)
  сбрасывают его в onRetune — сэмплы после ретюна спектрально разрывны

- IDevice::close() для чистого teardown из UI при закрытии окна
- RxWorker больше не вызывает stopStream (race с UI-потоком), это делает
  RxController на main thread после finished()

+++ DSP/BaseDemodHandler.h
@@ -46,7 +46,7 @@ public:
     [[nodiscard]] double ifRms() const { return dem_ ? dem_->ifRms() : 0.0; }
 
     // IPipelineHandler
-    void processBlock(const int16_t* iq, int count, double sampleRateHz) override;
+    void processBlock(const float* iq, int count, double sampleRateHz) override;
     void onStreamStarted(double sampleRateHz) override;
     void onStreamStopped() override;
